loadouts

+++ code/controllers/subsystem/rogue/role_class_handler/role_class_handler.dm
@@ -132,13 +132,23 @@ SUBSYSTEM_DEF(role_class_handler)
 			return FALSE // Along with stop here as they didn't get it.
 
 
-	H.advsetup = FALSE // This is actually on a lot of shit, so its a ghetto selector protector if u need one
+	if(related_handler && related_handler.linked_client)
+	{
+		related_handler.linked_client << browse(null, "window=class_handler_main")
+		related_handler.linked_client << browse(null, "window=class_select_yea")
+	}
+	var/loadout_pending = length(picked_class.loadout_options)
 	picked_class.equipme(H)
-	H.invisibility = 0
+	if(!loadout_pending)
+	{
+		H.advsetup = FALSE // keep the setup hud/blindness hint until loadout/equip completes
+		var/atom/movable/screen/advsetup/GET_IT_OUT = locate() in H.hud_used.static_inventory // dis line sux its basically a loop anyways if i remember
+		qdel(GET_IT_OUT)
+		H.cure_blind("advsetup")
+	}
+	if(H.invisibility == INVISIBILITY_MAXIMUM)
+		H.invisibility = 0
 	H.status_flags &= ~GODMODE
-	var/atom/movable/screen/advsetup/GET_IT_OUT = locate() in H.hud_used.static_inventory // dis line sux its basically a loop anyways if i remember
-	qdel(GET_IT_OUT)
-	H.cure_blind("advsetup")
 	H.remove_status_effect(/datum/status_effect/incapacitating/immobilized)
 
 	//If we get any plus factor at all, we run the datums boost proc on the human also.
